Move build_transactions_query into the library
